# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Commodity {
  symbol: string;
  name: string;
  unit: string;
  currency: string;
}

export interface PricePoint {
  date: string;
  close: number;
}

export interface MarketEvent {
  date: string;
  headline: string;
  category: string;
  sentimentScore: number;
  priceMove5d: number;
  source: string;
}

export interface DeskBrief {
  commodity: string;
  brief: string;
  latestPrice: number;
  changePct: number;
  asOf: string;
}

export type GetPricesParams = {
commodity: string;
};

export type GetEventsParams = {
commodity: string;
};

export type GetBriefParams = {
commodity: string;
};