# lib/api-zod/src/generated/types/commodity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Commodity {
  symbol: string;
  name: string;
  unit: string;
  currency: string;
}