# lib/api-zod/src/generated/types/deskBrief.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeskBrief {
  commodity: string;
  brief: string;
  latestPrice: number;
  changePct: number;
  asOf: string;
}