# lib/api-zod/src/generated/types/marketEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketEvent {
  date: string;
  headline: string;
  category: string;
  sentimentScore: number;
  priceMove5d: number;
  source: string;
}